requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
typing_extensions
tzdata